+++ src/fides/api/ops/service/saas_request/override_implementations/mailchimp_request_overrides.py
@@ -2,27 +2,24 @@ from json import dumps
 from typing import Any, Dict, List
 
 import pydash
-from requests import get, put
 
-from fides.api.ops.common_exceptions import (
-    ClientUnsuccessfulException,
-    ConnectionException,
-)
 from fides.api.ops.graph.traversal import TraversalNode
 from fides.api.ops.models.policy import Policy
 from fides.api.ops.models.privacy_request import PrivacyRequest
+from fides.api.ops.schemas.saas.shared_schemas import HTTPMethod, SaaSRequestParams
+from fides.api.ops.service.connectors.saas.authenticated_client import (
+    AuthenticatedClient,
+)
 from fides.api.ops.service.saas_request.saas_request_override_factory import (
     SaaSRequestType,
     register,
 )
 from fides.api.ops.util.collection_util import Row
-from fides.core.config import get_config
-
-CONFIG = get_config()
 
 
 @register("mailchimp_messages_access", [SaaSRequestType.READ])
 def mailchimp_messages_access(
+    client: AuthenticatedClient,
     node: TraversalNode,
     policy: Policy,
     privacy_request: PrivacyRequest,
@@ -60,26 +57,13 @@ def mailchimp_messages_access(
     processed_data = []
     if conversation_ids:
         for conversation_id in conversation_ids:
-            try:
-                response = get(
-                    url=f'https://{secrets["domain"]}/3.0/conversations/{conversation_id}/messages',
-                    auth=(secrets["username"], secrets["api_key"]),
-                )
 
-            # here we mimic the sort of error handling done in the core framework
-            # by the AuthenticatedClient. Extenders can chose to handle errors within
-            # their implementation as they wish.
-            except Exception as exc:  # pylint: disable=W0703
-                if CONFIG.dev_mode:  # pylint: disable=R1720
-                    raise ConnectionException(
-                        f"Operational Error connecting to Mailchimp API with error: {exc}"
-                    )
-                else:
-                    raise ConnectionException(
-                        "Operational Error connecting to MailchimpAPI."
-                    )
-            if not response.ok:
-                raise ClientUnsuccessfulException(status_code=response.status_code)
+            response = client.send(
+                SaaSRequestParams(
+                    method=HTTPMethod.GET,
+                    path=f"/3.0/conversations/{conversation_id}/messages",
+                )
+            )
 
             # unwrap and post-process response
             response_data = pydash.get(response.json(), "conversation_messages")
@@ -96,6 +80,7 @@ def mailchimp_messages_access(
 
 @register("mailchimp_member_update", [SaaSRequestType.UPDATE])
 def mailchimp_member_update(
+    client: AuthenticatedClient,
     param_values_per_row: List[Dict[str, Any]],
     policy: Policy,
     privacy_request: PrivacyRequest,
@@ -111,27 +96,14 @@ def mailchimp_member_update(
         # in this case, we can just put the masked object fields object
         # directly into the request body
         update_body = dumps(row_param_values["masked_object_fields"])
-        try:
-            response = put(
-                url=f'https://{secrets["domain"]}/3.0/lists/{list_id}/members/{subscriber_hash}',
-                auth=(secrets["username"], secrets["api_key"]),
-                data=update_body,
-            )
 
-        # here we mimic the sort of error handling done in the core framework
-        # by the AuthenticatedClient. Extenders can chose to handle errors within
-        # their implementation as they wish.
-        except Exception as e:
-            if CONFIG.dev_mode:  # pylint: disable=R1720
-                raise ConnectionException(
-                    f"Operational Error connecting to mailchimp API with error: {e}"
-                )
-            else:
-                raise ConnectionException(
-                    "Operational Error connecting to mailchimp API."
-                )
-        if not response.ok:
-            raise ClientUnsuccessfulException(status_code=response.status_code)
+        client.send(
+            SaaSRequestParams(
+                method=HTTPMethod.PUT,
+                path=f"/3.0/lists/{list_id}/members/{subscriber_hash}",
+                body=update_body,
+            )
+        )
 
         rows_updated += 1
     return rows_updated